+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.1] - 2020-06-17
+
+### Added
+
+- `getTarget()` method to access ShellSdk target object
+
 ## [1.5.0] - 2020-06-05
 
 ### Added
@@ -22,7 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Bugfix
 
-- Fix `.join` crash on  `GET_STORAGE_ITEM` event. from shell-host
+- Fix `.join` crash on `GET_STORAGE_ITEM` event. from shell-host
 
 ## [1.3.1] - 2020-05-11